move out json progress bar
